requirements.txt
anyio==4.3.0
asyncio==3.4.3
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.4
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
requests==2.31.0
setuptools==69.1.1
sniffio==1.3.1
Unidecode==1.3.8
urllib3==2.2.1
zope.interface==6.2
